Changes to animation thread to prevent a thread leak and to make sure animation doesn't stomp over a set color.
In StartRunningColors() it is possible for there to be a thread leak if it is
called multiple times without calling Stop() outside the method (and also
waiting). Stop() would just set _running to false and immediately after set
it back to true so a currently running animation thread most often wouldn't
be at the end of it's while loop at this time, so another thread will kick
off and the current one will continue. This causes odd animation behavior
and eventually crashes the Netduino. Because of a related race condition,
if an animation is stopped and a color is set right after, the color that
is set will most often get overriden by the animation thread turning off
the LED when it completes.